Cleanup and move the thumbnail wrapper
