+++ packages/story-editor/src/components/checklist/checks/elementLinkTappableRegionTooBig.js
@@ -35,7 +35,7 @@ import { LayerThumbnail, Thumbnail, THUMBNAIL_TYPES } from '../../thumbnail';
 import { filterStoryElements } from '../utils';
 import { useRegisterCheck } from '../countContext';
 import { useIsChecklistMounted } from '../popupMountedContext';
-import ThumbnailWrapper from '../../checklistCard/thumbnailWrapper';
+import { ThumbnailWrapper } from './shared';
 
 const MAX_LINK_SCREEN_PERCENT = 80;
