+++ packages/story-editor/src/components/checklist/checks/pageTooMuchText.js
@@ -38,7 +38,7 @@ import {
 } from '../utils';
 import { useRegisterCheck } from '../countContext';
 import { useIsChecklistMounted } from '../popupMountedContext';
-import ThumbnailWrapper from '../../checklistCard/thumbnailWrapper';
+import { ThumbnailWrapper } from './shared';
 
 /**
  * @typedef {import('../../../types').Page} Page
