+++ packages/story-editor/src/components/checklist/checks/textElementFontSizeTooSmall.js
@@ -35,7 +35,7 @@ import { LayerThumbnail, Thumbnail, THUMBNAIL_TYPES } from '../../thumbnail';
 import { filterStoryElements } from '../utils';
 import { useRegisterCheck } from '../countContext';
 import { useIsChecklistMounted } from '../popupMountedContext';
-import ThumbnailWrapper from '../../checklistCard/thumbnailWrapper';
+import { ThumbnailWrapper } from './shared';
 
 export function textElementFontSizeTooSmall(element) {
   return (
